requirements.txt
asgiref==3.2.7
Django==3.0.6
django-crispy-forms==1.9.1
django-tinymce==3.0.2
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
